Thought of a way to make the allcombinablepromotions methods a bit cleaner - Changes after initial submission
I didn't like how I was converting to and from sets and thought of a way it could be done with seq only.
I also think that these changes make the code more readable.

I believe this new solution should be more performant as well because it should be doing work on a smaller amount of subsets with the way i built the recursive method